First patch by square87 :
We compare: $user_name != [::abook::getNick $user]
to see if the name given from the server is different from the nick that we
stored. But [::abook::getNick $user] return a parsed nick... it means:
if you are using ColoredNicks, the original nick is [c=3]Square87[/c] but
using getnick you'll get Square87. This cause the useless calculation of
some code and in one case it causes that in CL we have a colored nick while
in topCW we have a nick without style. The solution is to compare $user_name
with the real nick stored:

$user_name != [::abook::getContactData $user nick]




git-svn-id: svn://svn.code.sf.net/p/amsn/code/trunk@9121 f248a3e1-9e13-0410-b2cf-a4429de6d898