First version of asynchronous host name resolver.
To make it work:

build utils/asyncresolver (edit Makefile to set paths, and run make).

It will be loaded automatically when amsn starts (at the end of amsncore.tcl). Socket command will be replaced, and host name resolution will be done 
asynchronously.

It would be great to integrate the build in the main Makefile...



git-svn-id: svn://svn.code.sf.net/p/amsn/code/trunk@10379 f248a3e1-9e13-0410-b2cf-a4429de6d898